requirements.txt
flask                         
requests                      
argparse                      
nltk                          
sentence-transformers     
torch                         
PyMuPDF       
faiss-cpu                
openai                      

